Extract EventListener class
This functionality is useful elsewhere.
